backend/requirements.txt
uvicorn
fastapi
pydantic
pandas
sentence_transformers
scikit-learn
google-cloud-bigquery
google-cloud-storage
db-dtypes